bot/requirements.txt
discord.py==1.2.3
pyyaml==5.1